Move snapshot logic into partition-store
