+++ crates/partition-store/src/snapshots/snapshot_task.rs
@@ -13,16 +13,15 @@ use std::time::SystemTime;
 
 use tracing::{debug, info, instrument, warn};
 
-use restate_partition_store::PartitionStoreManager;
-use restate_partition_store::snapshots::{
-    LocalPartitionSnapshot, PartitionSnapshotMetadata, SnapshotError, SnapshotErrorKind,
-    SnapshotFormatVersion,
-};
 use restate_types::identifiers::{PartitionId, SnapshotId};
 use restate_types::logs::Lsn;
 use restate_types::nodes_config::ClusterFingerprint;
 
-use crate::partition::snapshots::SnapshotRepository;
+use super::{
+    LocalPartitionSnapshot, PartitionSnapshotMetadata, SnapshotError, SnapshotErrorKind,
+    SnapshotFormatVersion, SnapshotRepository,
+};
+use crate::PartitionStoreManager;
 
 /// Creates a partition store snapshot along with Restate snapshot metadata.
 pub struct SnapshotPartitionTask {
