--- crates/worker/src/partition/snapshots/mod.rs
@@ -1,15 +1,0 @@
-// Copyright (c) 2023 - 2025 Restate Software, Inc., Restate GmbH.
-// All rights reserved.
-//
-// Use of this software is governed by the Business Source License
-// included in the LICENSE file.
-//
-// As of the Change Date specified in that file, in accordance with
-// the Business Source License, use of this software will be governed
-// by the Apache License, Version 2.0.
-
-mod repository;
-mod snapshot_task;
-
-pub use repository::SnapshotRepository;
-pub use snapshot_task::*;
